requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreschema==0.0.4
coverage==7.6.10
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-stubs==5.1.2
django-stubs-ext==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
